Fix compilation of sipXmediaAdapterLib tests under 64-bit Linux.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@10923 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXmediaAdapterLib/sipXmediaMediaProcessing/src/test/CpPhoneMediaInterfaceTest.cpp
@@ -1610,7 +1610,7 @@ class CpPhoneMediaInterfaceTest : public CppUnit::TestCase
         RTL_EVENT("Prompt", 1);
         printf("playing prompt in source 1\n");
         int waveHeaderSize = 116;
-        CPPUNIT_ASSERT_EQUAL((unsigned)160116, sizeof(sine_330hz_16b_8k_signed));
+        CPPUNIT_ASSERT_EQUAL((size_t)160116, sizeof(sine_330hz_16b_8k_signed));
         source1Interface->playBuffer((char*)&sine_330hz_16b_8k_signed[waveHeaderSize], 
                                      sizeof(sine_330hz_16b_8k_signed) - waveHeaderSize, 
                                      PROMPTBUF_SAMPLERATE,
@@ -1821,7 +1821,7 @@ class CpPhoneMediaInterfaceTest : public CppUnit::TestCase
 
         RTL_EVENT("Prompt", 1);
         int waveHeaderSize = 116;
-        CPPUNIT_ASSERT_EQUAL((unsigned)160116, sizeof(sine_330hz_16b_8k_signed));
+        CPPUNIT_ASSERT_EQUAL((size_t)160116, sizeof(sine_330hz_16b_8k_signed));
         mixedInterface->playBuffer((char*)&sine_330hz_16b_8k_signed[waveHeaderSize], 
                                    sizeof(sine_330hz_16b_8k_signed) - waveHeaderSize, 
                                    PROMPTBUF_SAMPLERATE,
